Move account selection logic to the plugin
