+++ accountsdb-plugin-manager/src/accountsdb_plugin_service.rs
@@ -1,6 +1,6 @@
 use {
     crate::{
-        accounts_update_notifier::{AccountsSelector, AccountsUpdateNotifierImpl},
+        accounts_update_notifier::AccountsUpdateNotifierImpl,
         accountsdb_plugin_manager::AccountsDbPluginManager,
         slot_status_observer::SlotStatusObserver,
     },
@@ -36,23 +36,9 @@ impl AccountsDbPluginService {
     ///    (.so file) to be loaded. The shared library must implement the `AccountsDbPlugin`
     ///    trait. And the shared library shall export a `C` function `_create_plugin` which
     ///    shall create the implementation of `AccountsDbPlugin` and returns to the caller.
-    ///    The `accounts_selector` section allows the user to controls accounts selections.
-    ///    "accounts_selector" : {
-    ///         "accounts" : \["pubkey-1", "pubkey-2", ..., "pubkey-n"\],
-    ///    }
-    ///    or:
-    ///    "accounts_selector" = {
-    ///         "owners" : \["pubkey-1', 'pubkey-2", ..., "pubkey-m"\]
-    ///    }
-    ///    When accounts and owners are specified, the program only filters only using accounts
-    ///    as it is more selective. And owners field is ignored. When only owners is specified,
-    ///    all accounts belonging to the owners will be streamed.
     ///    The rest of the JSON fields's definition is upto to the concrete plugin implementation
     ///    It is usually used to configure the connection information for the external data store.
-    ///    The accounts field support wildcard to select all accounts:
-    ///    "accounts_selector" : {
-    ///         "accounts" : \["*"\],
-    ///    }
+
     pub fn new(
         confirmed_bank_receiver: Receiver<BankNotification>,
         accountsdb_plugin_config_file: &Path,
@@ -69,11 +55,9 @@ impl AccountsDbPluginService {
         file.read_to_string(&mut contents).unwrap();
 
         let result: serde_json::Value = serde_json::from_str(&contents).unwrap();
-        let accounts_selector = Self::create_accounts_selector_from_config(&result);
 
         let accounts_update_notifier = Arc::new(RwLock::new(AccountsUpdateNotifierImpl::new(
             plugin_manager.clone(),
-            accounts_selector,
         )));
         let confirmed_slots_observer =
             SlotStatusObserver::new(confirmed_bank_receiver, accounts_update_notifier.clone());
@@ -97,38 +81,6 @@ impl AccountsDbPluginService {
         }
     }
 
-    fn create_accounts_selector_from_config(config: &serde_json::Value) -> AccountsSelector {
-        let accounts_selector = &config["accounts_selector"];
-
-        if accounts_selector.is_null() {
-            AccountsSelector::default()
-        } else {
-            let accounts = &accounts_selector["accounts"];
-            let accounts: Vec<String> = if accounts.is_array() {
-                accounts
-                    .as_array()
-                    .unwrap()
-                    .iter()
-                    .map(|val| val.as_str().unwrap().to_string())
-                    .collect()
-            } else {
-                Vec::default()
-            };
-            let owners = &accounts_selector["owners"];
-            let owners: Vec<String> = if owners.is_array() {
-                owners
-                    .as_array()
-                    .unwrap()
-                    .iter()
-                    .map(|val| val.as_str().unwrap().to_string())
-                    .collect()
-            } else {
-                Vec::default()
-            };
-            AccountsSelector::new(&accounts, &owners)
-        }
-    }
-
     pub fn get_accounts_update_notifier(&self) -> AccountsUpdateNotifier {
         self.accounts_update_notifier.clone()
     }
@@ -137,20 +89,5 @@ impl AccountsDbPluginService {
         self.confirmed_slots_observer.join()?;
         self.plugin_manager.write().unwrap().unload();
         Ok(())
-    }
-}
-
-#[cfg(test)]
-pub(crate) mod tests {
-    use {super::*, serde_json};
-
-    #[test]
-    fn test_accounts_selector_from_config() {
-        let config = "{\"accounts_selector\" : { \
-           \"owners\" : [\"9xQeWvG816bUx9EPjHmaT23yvVM2ZWbrrpZb9PusVFin\"] \
-        }}";
-
-        let config: serde_json::Value = serde_json::from_str(config).unwrap();
-        AccountsDbPluginService::create_accounts_selector_from_config(&config);
     }
 }
