+++ accountsdb-plugin-postgres/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod accounts_selector;
 pub mod accountsdb_plugin_postgres;
